fix: Move stripe's services under src/services
`@nimara/tooling` scans `src/services/*`, which the template already used, so
stripe's build had been looking in the wrong place. This is the rename that
closes that gap: the directory, the `@/apps/*` imports, the dev server's glob,
the eslint overrides that name client paths, and the README.

Both build targets verified: `node` writes `dist/handler/`, `vercel` writes
that plus the traced `index.js`.
