+++ apps/stripe/src/services/handler/client-entry-point.ts
@@ -5,7 +5,7 @@ import { createHtmlShell } from "@nimara/lib/hono/html-shell";
 
 import { container } from "@/container";
 
-// The build names the client bundle after the `src/apps/<name>` directory.
+// The build names the client bundle after the `src/services/<name>` directory.
 const APP_NAME = basename(dirname(fileURLToPath(import.meta.url)));
 
 const CONFIG = container.get("config");
