+++ apps/stripe/src/services/handler/client/views/app/components/config-details.tsx
